export from file functionality
